who knows what state this is in.
Summary:
Merge branch 'master' of bitbucket.org:scscastor/castor into leveldb-support

Merge branch 'master' of bitbucket.org:scscastor/castor into leveldb-support

cleaned up some naming stuff, no more rr_ prefix, no more strong references for pthreads stuff.

pthread_cond testbench.

Test Plan: unit test breaks currrently

Reviewers: ali

Differential Revision: https://phabricator.scs.stanford.edu/D29

+++ lib/Runtime/events.c
@@ -479,10 +479,10 @@ __rr_fcntl(int fd, int cmd, ...)
 {
     int result;
     va_list ap;
-    int arg;
+    int64_t arg;
 
     va_start(ap, cmd);
-    arg = va_arg(ap, int);
+    arg = va_arg(ap, int64_t);
     va_end(ap);
 
     switch (rrMode) {
